+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk: Added the `gmsol-sdk` crate: unlike the `gmsol` crate, this crate is built on top of the `gmsol-programs` crate and includes WASM support.
 - sdk: Added `get_token_accounts_by_owner_with_context` utility function.
 - sdk: Added `Client::rpc` method to access the shared `RpcClient`.
+- sdk: Added `WithSlot::slot_mut` and `WithSlot::value_mut` methods.
 - cli: Added the `treasury batch-withdraw` subcommand.
 - cli: Added `--authority` option for the `inspect price-feed` subcommand.
 - cli: Introduced a new ALT type `PriceFeed` for the `alt extend` subcommand.